chore: dryrun should not depend on run.service
First step of the refactoring of run.service. The end goal is to be able
to split `run` into two components: one to start a script execution and
the other one to complete/report on it results.
In order to be able to do that I am first removing the usage of
run.service by cli dryrun. Very little logic from the run.service is
actually necessary for dryrun and it is actually ok for dryrun execution
to be synchronous since they are executed locally via the CLI
I have not touched the runScript interface, other than some input fields
tweaking. Even though I could push the refactoring further and avoid
nangProps attributes that are only required for dryrun (ex: logMessages
for instance I believe) I don't want to dig into a rabbit hole too much
for now.